Add ARM template for Azure App Service deployment
This commit introduces a new Azure Resource Manager (ARM) template in JSON format. The template defines a deployment for an Azure App Service and its associated resources, including parameters for resource group name, location, and resource name. It also specifies the creation of a resource group and includes configurations for an App Service and an App Service Plan, detailing identity, site settings, and SKU information.